+++ packages/core/src/utils/generate-response.ts
@@ -1,6 +1,6 @@
-import type { Response } from 'express';
-import { getReasonPhrase, StatusCodes } from 'http-status-codes';
-import { GenerateResponse } from '@dry-express-responses/types';
+import type {Response} from 'express';
+import {getReasonPhrase, StatusCodes} from 'http-status-codes';
+import {GenerateResponse} from '@dry-express-responses/types';
 
 /**
  * Reduces the use of res.status() and res.send(), standardizes the response with a consistent format in addition to TypeScript typing, while allowing data and a message to be passed in.
@@ -9,13 +9,13 @@ import { GenerateResponse } from '@dry-express-responses/types';
  */
 export const generateResponse =
 	<TStatus extends StatusCodes>(res: Response, status: TStatus) =>
-	// Only has errors if status is not OK or CREATED.
-	<TData>(
-		payload:
-			| GenerateResponse<TStatus, TData>
-			| Record<never, never> = {},
-	) =>
-		res.status(status).send({
-			status: getReasonPhrase(status),
-			...payload,
-		});
+		// Only has errors if status is not OK or CREATED.
+		<TData, TPayload extends Record<PropertyKey, any>>(
+			payload:
+				| GenerateResponse<TStatus, TData, TPayload>
+				| Record<never, never> = {},
+		) =>
+			res.status(status).send({
+				status: getReasonPhrase(status),
+				...payload,
+			});