+++ packages/zod-request-validation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @dry-express-responses/zod-request-validation
 
+
 ## 2.0.11
 
 ### Patch Changes
@@ -12,11 +13,11 @@
 
 ### Patch Changes
 
-- 478a56b: testing
 - 106fd71: added a responses factory and updated the readme.md examples
 - bcdfd3a: added a responses factory and updated examples in README.md
 - Updated dependencies [478a56b]
 - Updated dependencies [106fd71]
 - Updated dependencies [bcdfd3a]
-  - @dry-express-responses/errors@0.2.10
+- Updated dependencies [841088f]
+  - @dry-express-responses/errors@1.0.0
   - @dry-express-responses/zod@0.0.41